refactor: extract platform address transitions into PlatformAddressService on DIP-17 derivation (sdk 1.5.0-dev.5)
