+++ package.json
@@ -36,7 +36,7 @@
     "classic-level": "^3.0.0",
     "dash-core-p2p": "https://github.com/pshenmic/dash-core-p2p#0121cf0eb934ff7378cf7aa42ec13f2907a7eb50",
     "dash-core-sdk": "1.1.3-dev.4",
-    "dash-platform-sdk": "1.5.0-dev.4",
+    "dash-platform-sdk": "1.5.0-dev.5",
     "dash-ui-kit": "1.0.94",
     "knex": "^3.1.0",
     "react-qr-code": "^2.0.18",
