+++ src/main/src/api/wallet/sendIdentityCredits.ts
@@ -1,12 +1,12 @@
 import { IpcMainInvokeEvent } from 'electron/utility'
-import { WalletService } from '../../services/WalletService'
+import { PlatformAddressService } from '../../services/PlatformAddressService'
 import { PlatformSendResult } from '../../types/PlatformSendResult'
 
 export class SendIdentityCreditsHandler {
-  private walletService: WalletService
+  private platformAddressService: PlatformAddressService
 
-  constructor(walletService: WalletService) {
-    this.walletService = walletService
+  constructor(platformAddressService: PlatformAddressService) {
+    this.platformAddressService = platformAddressService
   }
 
   handle = async (
@@ -17,7 +17,7 @@ export class SendIdentityCreditsHandler {
     amountCredits: string,
     password: string,
   ): Promise<PlatformSendResult> => {
-    return this.walletService.sendIdentityCreditsToAddresses(
+    return this.platformAddressService.sendIdentityCreditsToAddresses(
       walletId,
       identityId,
       [{address: toAddress, amountCredits: BigInt(amountCredits)}],
