+++ src/main/src/database/WalletDAO.ts
@@ -1,8 +1,8 @@
 import {Wallet} from '../types/Wallet'
 import {QueryStatus} from "../types/QueryStatus";
 
-function fromRow({wallet_id, label, network, encrypted_mnemonic, selected}): Wallet {
-  return {walletId: wallet_id, network, label, encryptedMnemonic: encrypted_mnemonic, selected: Boolean(selected)}
+function fromRow({wallet_id, label, network, encrypted_mnemonic, selected, platform_xpub}): Wallet {
+  return {walletId: wallet_id, network, label, encryptedMnemonic: encrypted_mnemonic, selected: Boolean(selected), platformXpub: platform_xpub ?? null}
 }
 
 export class WalletDAO {
@@ -44,7 +44,7 @@ export class WalletDAO {
 
   getWalletById = async (walletId): Promise<Wallet | null> => {
     const rows = await this.knex('wallet')
-      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected')
+      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected', 'platform_xpub')
       .where('wallet_id', walletId)
       .limit(1)
 
@@ -59,14 +59,14 @@ export class WalletDAO {
 
   getAllWallets = async (): Promise<Wallet[]> => {
     const rows = await this.knex('wallet')
-      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected')
+      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected', 'platform_xpub')
 
     return rows.map(fromRow)
   }
 
   getSelectedWallet = async (): Promise<Wallet | null> => {
     const rows = await this.knex('wallet')
-      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected')
+      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected', 'platform_xpub')
       .where('selected', true)
       .limit(1)
 
@@ -99,6 +99,12 @@ export class WalletDAO {
     }
   }
 
+  setPlatformXpub = async (walletId: string, platformXpub: string): Promise<void> => {
+    await this.knex('wallet')
+      .update({platform_xpub: platformXpub})
+      .where('wallet_id', walletId)
+  }
+
   updateLabel = async (walletId: string, label: string | null): Promise<QueryStatus> => {
     try {
       const result = await this.knex('wallet')
@@ -128,7 +134,7 @@ export class WalletDAO {
 
   getWalletsByNetwork = async (network): Promise<Wallet[]> => {
     const rows = await this.knex('wallet')
-      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected')
+      .select('encrypted_mnemonic', 'network', 'wallet_id', 'label', 'selected', 'platform_xpub')
       .where('network', network)
 
     return rows.map(fromRow)
