+++ src/main/src/services/WalletService.ts
@@ -14,19 +14,7 @@ import {Address} from '../types/Address'
 import {GroupedAddresses} from '../types/GroupedAddresses'
 import {Identity, IdentityInfo} from '../types/Identity'
 import {Wallet} from '../types/Wallet'
-import {
-  InputAddressWASM,
-  OutputAddressWASM,
-  AddressFundsFeeStrategyStepWASM,
-  AddressWitnessWASM,
-  AddressFundsTransferTransitionWASM,
-  IdentityTopUpFromAddressesTransitionWASM,
-  AddressCreditWithdrawalTransitionWASM,
-  IdentityCreateFromAddressesTransitionWASM,
-  IdentityPublicKeyInCreationWASM,
-  PrivateKeyWASM,
-  IdentityPublicKeyWASM,
-} from 'dash-platform-sdk/types.js'
+import {PrivateKeyWASM} from 'dash-platform-sdk/types.js'
 import {BlockJSON} from "dash-core-sdk/src/types";
 import {QueryStatus} from "../types/QueryStatus";
 import {WalletBalance} from "../types/WalletBalance";
@@ -36,33 +24,11 @@ import {selectCoins, SelectableUtxo} from "../utils/coinSelection";
 import {dedupeTransactions} from "../utils/dedupeTransactions";
 import {CoreTransactionService, TransferInput} from "./CoreTransactionService";
 import {decryptMnemonic, encryptMnemonic} from "../utils";
-import {coreAddressToPlatformAddress} from "./platformAddress";
-import {PlatformAddressEntry} from "../types/PlatformAddress";
-import {ShieldedMemoWASM} from 'pshenmic-dpp'
-import {ShieldResult} from "../types/ShieldResult";
-import {PlatformSendResult} from "../types/PlatformSendResult";
-import {IdentityCreateResult} from "../types/IdentityCreateResult";
-import {
-  PlatformSourceCandidate,
-  selectPlatformSource,
-  selectPlatformInputsWithFee,
-  topUpFeeCredits,
-  TRANSFER_FEE_CREDITS,
-  WITHDRAWAL_FEE_CREDITS,
-  CORE_FEE_PER_BYTE,
-  MAX_RECIPIENTS,
-  MIN_OUTPUT_CREDITS,
-  identityTransferFeeCredits,
-  identityCreateFeeCredits,
-} from "../utils/platformTransfer";
-import {coreAddressToScript} from "../utils/coreScript";
-import {matchIdentityKey, DerivedKeyHash} from "../utils/identityKeys";
 
 const ADDRESS_LOOKAHEAD = 20
 const IDENTITY_LOOKAHEAD = 10
-const IDENTITY_KEY_LOOKAHEAD = 20
 const COIN_TYPE: Record<Network, number> = {mainnet: 5, testnet: 1}
-const SHIELDED_ACCOUNT = 0
+const PLATFORM_ACCOUNT = 0
 
 export class WalletService {
   private walletDAO: WalletDAO
@@ -128,6 +94,9 @@ export class WalletService {
     const hdKey = this.sdk.keyPair.seedToHdKey(seed, network)
     const coinType = COIN_TYPE[network]
     const accountId = 0
+
+    const platformXpub = await this.sdk.keyPair.derivePlatformAccountXpub(seed, network, PLATFORM_ACCOUNT)
+    await this.walletDAO.setPlatformXpub(walletId, platformXpub)
 
     const addresses: Address[] = []
 
@@ -269,7 +238,14 @@ export class WalletService {
 
     const address = this.sdk.keyPair.p2pkhAddress(key.publicKey, wallet.network)
 
-    return address === referenceWalletAddress.address;
+    const isValid = address === referenceWalletAddress.address
+
+    if (isValid && wallet.platformXpub == null) {
+      const platformXpub = await this.sdk.keyPair.derivePlatformAccountXpub(seed, wallet.network, PLATFORM_ACCOUNT)
+      await this.walletDAO.setPlatformXpub(walletId, platformXpub)
+    }
+
+    return isValid
   }
 
   async setAddressLabel(walletId: string, address: string, label: string): Promise<QueryStatus> {
@@ -631,604 +607,6 @@ export class WalletService {
 
   async getIdentityNonce(identifier: string): Promise<bigint> {
     return this.sdk.identities.getIdentityNonce(identifier)
-  }
-
-  async getPlatformAddresses(walletId: string): Promise<PlatformAddressEntry[]> {
-    const wallet = await this.walletDAO.getWalletById(walletId)
-
-    if (wallet == null) {
-      throw new Error('Wallet not found')
-    }
-
-    this.sdk.setNetwork(wallet.network)
-
-    const candidates = await this.loadPlatformCandidates(walletId, wallet.network)
-    return candidates.map(candidate => ({
-      platformAddress: candidate.platformAddress,
-      balanceCredits: candidate.balanceCredits.toString(),
-      nonce: candidate.nonce,
-    }))
-  }
-
-  private async loadPlatformCandidates(walletId: string, network: Network): Promise<PlatformSourceCandidate[]> {
-    const owned = await this.buildPlatformOwned(walletId, network)
-    const infoByPlatformAddress = await this.fetchPlatformAddressInfos(
-      owned.map(entry => entry.platformAddress),
-      network,
-    )
-    return owned.map(entry => {
-      const info = infoByPlatformAddress.get(entry.platformAddress)
-      return {
-        ...entry,
-        balanceCredits: info?.balance ?? 0n,
-        nonce: info?.nonce ?? 0,
-      }
-    })
-  }
-
-  private async deriveAddressPrivateKey(
-    hdKey: ReturnType<DashPlatformSDK['keyPair']['seedToHdKey']>,
-    source: {derivationPath: string; coreAddress: string},
-    network: Network,
-  ): Promise<PrivateKeyWASM> {
-    const derived = await this.sdk.keyPair.derivePath(hdKey, source.derivationPath)
-    if (!derived.privateKey) {
-      throw new Error(`Failed to derive private key for ${source.coreAddress}`)
-    }
-    return PrivateKeyWASM.fromBytes(derived.privateKey as Uint8Array, network)
-  }
-
-  private async signAddressInputs(
-    signable: Uint8Array,
-    sources: Array<{derivationPath: string; coreAddress: string}>,
-    hdKey: ReturnType<DashPlatformSDK['keyPair']['seedToHdKey']>,
-    network: Network,
-  ): Promise<AddressWitnessWASM[]> {
-    const witnesses: AddressWitnessWASM[] = []
-    for (const source of sources) {
-      const privateKey = await this.deriveAddressPrivateKey(hdKey, source, network)
-      witnesses.push(AddressWitnessWASM.P2PKH(privateKey.sign(signable)))
-    }
-    return witnesses
-  }
-
-  private async buildPlatformOwned(
-    walletId: string,
-    network: Network,
-  ): Promise<Array<{platformAddress: string; coreAddress: string; derivationPath: string}>> {
-    const grouped = await this.addressDAO.getAddressesByWalletId(walletId)
-    return [...grouped.receiving, ...grouped.change].map(address => ({
-      platformAddress: coreAddressToPlatformAddress(address.address, network),
-      coreAddress: address.address,
-      derivationPath: address.derivationPath,
-    }))
-  }
-
-  private async fetchPlatformAddressInfos(platformAddresses: string[], network: Network): Promise<Map<string, { balance: bigint; nonce: number }>> {
-    const result = new Map<string, { balance: bigint; nonce: number }>()
-
-    if (platformAddresses.length === 0) {
-      return result
-    }
-
-    try {
-      const infos = await this.sdk.platformAddresses.getAddressesInfos(platformAddresses)
-      for (const info of infos) {
-        result.set(info.address.toBech32m(network), { balance: info.balance, nonce: info.nonce })
-      }
-      return result
-    } catch {
-      const settled = await Promise.allSettled(
-        platformAddresses.map(address => this.sdk.platformAddresses.getAddressInfo(address))
-      )
-      settled.forEach((outcome, i) => {
-        if (outcome.status === 'fulfilled') {
-          result.set(platformAddresses[i], { balance: outcome.value.balance, nonce: outcome.value.nonce })
-        }
-      })
-      return result
-    }
-  }
-
-  async sendPlatformTransfer(
-    walletId: string,
-    fromPlatformAddress: string,
-    toPlatformAddress: string,
-    amountCredits: bigint,
-    password: string,
-  ): Promise<PlatformSendResult> {
-    if (amountCredits <= 0n) {
-      throw new Error('Send amount must be greater than zero')
-    }
-
-    const wallet = await this.walletDAO.getWalletById(walletId)
-    if (wallet == null) {
-      throw new Error('Wallet not found')
-    }
-    const network = wallet.network
-
-    this.sdk.setNetwork(network)
-
-    let decryptedMnemonic: string
-    try {
-      decryptedMnemonic = decryptMnemonic(wallet.encryptedMnemonic, password)
-    } catch {
-      throw new Error('Invalid wallet password')
-    }
-
-    const candidates = await this.loadPlatformCandidates(walletId, network)
-
-    const source = selectPlatformSource(candidates, amountCredits, fromPlatformAddress || undefined)
-
-    if (toPlatformAddress === source.platformAddress) {
-      throw new Error('Recipient must be different from the source address')
-    }
-
-    const seed = this.sdk.keyPair.mnemonicToSeed(decryptedMnemonic)
-    const hdKey = this.sdk.keyPair.seedToHdKey(seed, network)
-
-    const inputs = [new InputAddressWASM(source.platformAddress, source.nonce + 1, amountCredits)]
-    const outputs = [new OutputAddressWASM(toPlatformAddress, amountCredits)]
-    const feeStrategy = [AddressFundsFeeStrategyStepWASM.DeductFromInput(0)]
-
-    const unsignedSt = this.sdk.platformAddresses.createStateTransition('addressFundsTransfer', {
-      inputs,
-      feeStrategy,
-      userFeeIncrease: 0,
-      inputWitness: [],
-      outputs,
-    })
-
-    const signable = unsignedSt.getSignableBytes()
-    const witnesses = await this.signAddressInputs(signable, [source], hdKey, network)
-
-    const transition = AddressFundsTransferTransitionWASM.fromStateTransition(unsignedSt)
-    transition.inputWitness = witnesses
-    const signedSt = transition.toStateTransition()
-
-    await this.sdk.stateTransitions.broadcast(signedSt)
-    await this.sdk.stateTransitions.waitForStateTransitionResult(signedSt)
-
-    return {
-      stHash: signedSt.hash(false),
-      amountCredits: amountCredits.toString(),
-      feeCredits: TRANSFER_FEE_CREDITS.toString(),
-      fromAddress: source.platformAddress,
-      toAddress: toPlatformAddress,
-    }
-  }
-
-  private async resolveIdentitySigningKey(
-    identity: Identity,
-    hdKey: ReturnType<DashPlatformSDK['keyPair']['seedToHdKey']>,
-    network: Network,
-  ): Promise<{privateKey: PrivateKeyWASM; publicKey: IdentityPublicKeyWASM}> {
-    const identityKeys = await this.sdk.identities.getIdentityPublicKeys(identity.identifier)
-
-    const derivedKeys: Array<{keyIndex: number; privateKey: PrivateKeyWASM}> = []
-    const derivedHashes: DerivedKeyHash[] = []
-    for (let keyIndex = 0; keyIndex < IDENTITY_KEY_LOOKAHEAD; keyIndex++) {
-      const child = this.sdk.keyPair.deriveIdentityPrivateKey(hdKey, identity.identityIndex, keyIndex, network)
-      if (!child.privateKey) continue
-      const privateKey = PrivateKeyWASM.fromBytes(child.privateKey as Uint8Array, network)
-      derivedKeys.push({keyIndex, privateKey})
-      derivedHashes.push({keyIndex, publicKeyHashHex: privateKey.getPublicKeyHash()})
-    }
-
-    const match = matchIdentityKey(
-      identityKeys.map(key => ({
-        keyId: key.keyId,
-        purpose: key.purpose,
-        publicKeyHashHex: key.getPublicKeyHash(),
-      })),
-      derivedHashes,
-    )
-    if (match == null) {
-      throw new Error('This identity has no transfer key this wallet can sign with')
-    }
-
-    const derived = derivedKeys.find(entry => entry.keyIndex === match.keyIndex)
-    const publicKey = identityKeys.find(key => key.keyId === match.keyId)
-    if (derived == null || publicKey == null) {
-      throw new Error('This identity has no transfer key this wallet can sign with')
-    }
-
-    return {privateKey: derived.privateKey, publicKey}
-  }
-
-  async sendIdentityCreditsToAddresses(
-    walletId: string,
-    identityIdentifier: string,
-    recipients: Array<{address: string; amountCredits: bigint}>,
-    password: string,
-  ): Promise<PlatformSendResult> {
-    if (recipients.length === 0 || recipients.length > MAX_RECIPIENTS) {
-      throw new Error(`Recipient count must be between 1 and ${MAX_RECIPIENTS}`)
-    }
-    for (const recipient of recipients) {
-      if (recipient.amountCredits < MIN_OUTPUT_CREDITS) {
-        throw new Error(`Minimum amount per recipient is ${MIN_OUTPUT_CREDITS.toString()} credits`)
-      }
-    }
-
-    const wallet = await this.walletDAO.getWalletById(walletId)
-    if (wallet == null) {
-      throw new Error('Wallet not found')
-    }
-    const network = wallet.network
-
-    this.sdk.setNetwork(network)
-
-    let decryptedMnemonic: string
-    try {
-      decryptedMnemonic = decryptMnemonic(wallet.encryptedMnemonic, password)
-    } catch {
-      throw new Error('Invalid wallet password')
-    }
-
-    const identities = await this.identityDAO.getIdentitiesByWalletId(walletId)
-    const identity = identities.find(entry => entry.identifier === identityIdentifier)
-    if (identity == null) {
-      throw new Error('Identity not found in this wallet')
-    }
-
-    const totalCredits = recipients.reduce((sum, recipient) => sum + recipient.amountCredits, 0n)
-    const feeCredits = identityTransferFeeCredits(recipients.length)
-
-    const balance = await this.sdk.identities.getIdentityBalance(identityIdentifier)
-    if (balance < totalCredits + feeCredits) {
-      throw new Error('Identity has insufficient credits for this transfer plus fee')
-    }
-
-    const seed = this.sdk.keyPair.mnemonicToSeed(decryptedMnemonic)
-    const hdKey = this.sdk.keyPair.seedToHdKey(seed, network)
-    const {privateKey, publicKey} = await this.resolveIdentitySigningKey(identity, hdKey, network)
-
-    const nonce = await this.sdk.identities.getIdentityNonce(identityIdentifier) + 1n
-
-    const unsignedSt = this.sdk.platformAddresses.createStateTransition('identityCreditTransferToAddresses', {
-      identityId: identityIdentifier,
-      recipients: recipients.map(recipient => new OutputAddressWASM(recipient.address, recipient.amountCredits)),
-      nonce,
-      userFeeIncrease: 0,
-    })
-
-    const signature = unsignedSt.sign(privateKey, publicKey)
-    if (unsignedSt.signature == null || unsignedSt.signature.length === 0) {
-      unsignedSt.signature = signature
-      unsignedSt.signaturePublicKeyId = publicKey.keyId
-    }
-
-    await this.sdk.stateTransitions.broadcast(unsignedSt)
-    await this.sdk.stateTransitions.waitForStateTransitionResult(unsignedSt)
-
-    return {
-      stHash: unsignedSt.hash(false),
-      amountCredits: totalCredits.toString(),
-      feeCredits: feeCredits.toString(),
-      fromAddress: identityIdentifier,
-      toAddress: recipients[0].address,
-    }
-  }
-
-  async createIdentityFromAddresses(
-    walletId: string,
-    fromPlatformAddress: string | null,
-    amountCredits: bigint,
-    password: string,
-  ): Promise<IdentityCreateResult> {
-    if (amountCredits < MIN_OUTPUT_CREDITS) {
-      throw new Error(`Minimum identity funding is ${MIN_OUTPUT_CREDITS.toString()} credits`)
-    }
-
-    const wallet = await this.walletDAO.getWalletById(walletId)
-    if (wallet == null) {
-      throw new Error('Wallet not found')
-    }
-    const network = wallet.network
-
-    this.sdk.setNetwork(network)
-
-    let decryptedMnemonic: string
-    try {
-      decryptedMnemonic = decryptMnemonic(wallet.encryptedMnemonic, password)
-    } catch {
-      throw new Error('Invalid wallet password')
-    }
-
-    const existing = await this.identityDAO.getIdentitiesByWalletId(walletId)
-    const identityIndex = existing.reduce((max, identity) => Math.max(max, identity.identityIndex), -1) + 1
-
-    const seed = this.sdk.keyPair.mnemonicToSeed(decryptedMnemonic)
-    const hdKey = this.sdk.keyPair.seedToHdKey(seed, network)
-
-    const keySpecs: Array<{purpose: 'AUTHENTICATION' | 'TRANSFER'; securityLevel: 'MASTER' | 'HIGH' | 'CRITICAL'}> = [
-      {purpose: 'AUTHENTICATION', securityLevel: 'MASTER'},
-      {purpose: 'AUTHENTICATION', securityLevel: 'HIGH'},
-      {purpose: 'AUTHENTICATION', securityLevel: 'CRITICAL'},
-      {purpose: 'TRANSFER', securityLevel: 'CRITICAL'},
-    ]
-
-    const identityKeys = keySpecs.map((spec, keyIndex) => {
-      const child = this.sdk.keyPair.deriveIdentityPrivateKey(hdKey, identityIndex, keyIndex, network)
-      if (!child.privateKey || !child.publicKey) {
-        throw new Error(`Failed to derive identity key at index ${keyIndex}`)
-      }
-      return {
-        keyId: keyIndex,
-        spec,
-        privateKey: PrivateKeyWASM.fromBytes(child.privateKey as Uint8Array, network),
-        publicKey: child.publicKey as Uint8Array,
-      }
-    })
-
-    const candidates = await this.loadPlatformCandidates(walletId, network)
-    const plan = selectPlatformInputsWithFee(
-      candidates,
-      amountCredits,
-      () => identityCreateFeeCredits(keySpecs.length),
-      fromPlatformAddress ?? undefined,
-    )
-
-    const inputs = plan.inputs.map(({candidate, credits}) =>
-      new InputAddressWASM(candidate.platformAddress, candidate.nonce + 1, credits))
-    const feeStrategy = [AddressFundsFeeStrategyStepWASM.DeductFromInput(0)]
-
-    const publicKeysInCreation = identityKeys.map(key =>
-      new IdentityPublicKeyInCreationWASM(key.keyId, key.spec.purpose, key.spec.securityLevel, 'ECDSA_SECP256K1', false, key.publicKey))
-
-    const unsignedSt = this.sdk.platformAddresses.createStateTransition('identityCreateFromAddresses', {
-      publicKeys: publicKeysInCreation,
-      inputs,
-      feeStrategy,
-      inputWitness: [],
-      userFeeIncrease: 0,
-    })
-
-    const signable = unsignedSt.getSignableBytes()
-
-    const signedKeys = identityKeys.map(key =>
-      new IdentityPublicKeyInCreationWASM(key.keyId, key.spec.purpose, key.spec.securityLevel, 'ECDSA_SECP256K1', false, key.publicKey, key.privateKey.sign(signable)))
-    const witnesses = await this.signAddressInputs(signable, plan.inputs.map(input => input.candidate), hdKey, network)
-
-    const transition = IdentityCreateFromAddressesTransitionWASM.fromStateTransition(unsignedSt)
-    transition.publicKeys = signedKeys
-    transition.inputWitness = witnesses
-    const signedSt = transition.toStateTransition()
-
-    await this.sdk.stateTransitions.broadcast(signedSt)
-    await this.sdk.stateTransitions.waitForStateTransitionResult(signedSt)
-
-    const masterKeyHash = identityKeys[0].privateKey.getPublicKeyHash()
-    let identifier: string | null = null
-    try {
-      const identity = await this.sdk.identities.getIdentityByPublicKeyHash(masterKeyHash)
-      identifier = identity.id.base58()
-    } catch {
-      try {
-        const identity = await this.sdk.identities.getIdentityByNonUniquePublicKeyHash(masterKeyHash)
-        identifier = identity.id.base58()
-      } catch {
-        throw new Error('Identity was broadcast but could not be resolved yet — re-open the wallet to pick it up')
-      }
-    }
-
-    await this.identityDAO.insertIdentities([{
-      walletId,
-      identityIndex,
-      derivationPath: `m/9'/${COIN_TYPE[network]}'/0'/0/${identityIndex}`,
-      identifier,
-    }])
-
-    return {
-      identifier,
-      identityIndex,
-      stHash: signedSt.hash(false),
-      amountCredits: amountCredits.toString(),
-      feeCredits: plan.feeCredits.toString(),
-      fromAddress: plan.inputs[0].candidate.platformAddress,
-    }
-  }
-
-  async topUpIdentityFromAddresses(
-    walletId: string,
-    identityId: string,
-    fromPlatformAddress: string | null,
-    amountCredits: bigint,
-    password: string,
-  ): Promise<PlatformSendResult> {
-    if (amountCredits <= 0n) {
-      throw new Error('Top-up amount must be greater than zero')
-    }
-
-    const wallet = await this.walletDAO.getWalletById(walletId)
-    if (wallet == null) {
-      throw new Error('Wallet not found')
-    }
-    const network = wallet.network
-
-    this.sdk.setNetwork(network)
-
-    let decryptedMnemonic: string
-    try {
-      decryptedMnemonic = decryptMnemonic(wallet.encryptedMnemonic, password)
-    } catch {
-      throw new Error('Invalid wallet password')
-    }
-
-    try {
-      await this.sdk.identities.getIdentityByIdentifier(identityId)
-    } catch {
-      throw new Error('Identity not found on Platform')
-    }
-
-    const candidates = await this.loadPlatformCandidates(walletId, network)
-    const plan = selectPlatformInputsWithFee(candidates, amountCredits, topUpFeeCredits, fromPlatformAddress ?? undefined)
-
-    const seed = this.sdk.keyPair.mnemonicToSeed(decryptedMnemonic)
-    const hdKey = this.sdk.keyPair.seedToHdKey(seed, network)
-
-    const inputs = plan.inputs.map(({candidate, credits}) =>
-      new InputAddressWASM(candidate.platformAddress, candidate.nonce + 1, credits))
-    const feeStrategy = [AddressFundsFeeStrategyStepWASM.DeductFromInput(0)]
-
-    const unsignedSt = this.sdk.platformAddresses.createStateTransition('identityTopUpFromAddresses', {
-      identityId,
-      inputs,
-      feeStrategy,
-      inputWitness: [],
-      userFeeIncrease: 0,
-    })
-
-    const signable = unsignedSt.getSignableBytes()
-    const witnesses = await this.signAddressInputs(signable, plan.inputs.map(input => input.candidate), hdKey, network)
-
-    const transition = IdentityTopUpFromAddressesTransitionWASM.fromStateTransition(unsignedSt)
-    transition.inputWitness = witnesses
-    const signedSt = transition.toStateTransition()
-
-    await this.sdk.stateTransitions.broadcast(signedSt)
-    await this.sdk.stateTransitions.waitForStateTransitionResult(signedSt)
-
-    return {
-      stHash: signedSt.hash(false),
-      amountCredits: amountCredits.toString(),
-      feeCredits: plan.feeCredits.toString(),
-      fromAddress: plan.inputs[0].candidate.platformAddress,
-      toAddress: identityId,
-    }
-  }
-
-  async withdrawPlatformToCore(
-    walletId: string,
-    fromPlatformAddress: string | null,
-    toCoreAddress: string,
-    amountCredits: bigint,
-    password: string,
-  ): Promise<PlatformSendResult> {
-    if (amountCredits <= 0n) {
-      throw new Error('Withdrawal amount must be greater than zero')
-    }
-
-    const wallet = await this.walletDAO.getWalletById(walletId)
-    if (wallet == null) {
-      throw new Error('Wallet not found')
-    }
-    const network = wallet.network
-
-    this.sdk.setNetwork(network)
-
-    const outputScript = coreAddressToScript(toCoreAddress, network)
-
-    let decryptedMnemonic: string
-    try {
-      decryptedMnemonic = decryptMnemonic(wallet.encryptedMnemonic, password)
-    } catch {
-      throw new Error('Invalid wallet password')
-    }
-
-    const candidates = await this.loadPlatformCandidates(walletId, network)
-    const plan = selectPlatformInputsWithFee(
-      candidates,
-      amountCredits,
-      () => WITHDRAWAL_FEE_CREDITS,
-      fromPlatformAddress ?? undefined,
-    )
-
-    const seed = this.sdk.keyPair.mnemonicToSeed(decryptedMnemonic)
-    const hdKey = this.sdk.keyPair.seedToHdKey(seed, network)
-
-    const inputs = plan.inputs.map(({candidate, credits}) =>
-      new InputAddressWASM(candidate.platformAddress, candidate.nonce + 1, credits))
-    const feeStrategy = [AddressFundsFeeStrategyStepWASM.DeductFromInput(0)]
-
-    const unsignedSt = this.sdk.platformAddresses.createStateTransition('addressCreditWithdrawal', {
-      inputs,
-      feeStrategy,
-      inputWitness: [],
-      userFeeIncrease: 0,
-      coreFeePerByte: CORE_FEE_PER_BYTE,
-      pooling: 'Never',
-      outputScript,
-    })
-
-    const signable = unsignedSt.getSignableBytes()
-    const witnesses = await this.signAddressInputs(signable, plan.inputs.map(input => input.candidate), hdKey, network)
-
-    const transition = AddressCreditWithdrawalTransitionWASM.fromStateTransition(unsignedSt)
-    transition.inputWitness = witnesses
-    const signedSt = transition.toStateTransition()
-
-    await this.sdk.stateTransitions.broadcast(signedSt)
-    await this.sdk.stateTransitions.waitForStateTransitionResult(signedSt)
-
-    return {
-      stHash: signedSt.hash(false),
-      amountCredits: amountCredits.toString(),
-      feeCredits: plan.feeCredits.toString(),
-      fromAddress: plan.inputs[0].candidate.platformAddress,
-      toAddress: toCoreAddress,
-    }
-  }
-
-  async shieldToPool(
-    walletId: string,
-    fromPlatformAddress: string,
-    amountCredits: bigint,
-    password: string,
-  ): Promise<ShieldResult> {
-    if (amountCredits <= 0n) {
-      throw new Error('Shield amount must be greater than zero')
-    }
-
-    const wallet = await this.walletDAO.getWalletById(walletId)
-    if (wallet == null) {
-      throw new Error('Wallet not found')
-    }
-    const network = wallet.network
-
-    this.sdk.setNetwork(network)
-
-    let decryptedMnemonic: string
-    try {
-      decryptedMnemonic = decryptMnemonic(wallet.encryptedMnemonic, password)
-    } catch {
-      throw new Error('Invalid wallet password')
-    }
-
-    const candidates = await this.loadPlatformCandidates(walletId, network)
-
-    const source = selectPlatformSource(candidates, amountCredits, fromPlatformAddress || undefined)
-
-    const seed = this.sdk.keyPair.mnemonicToSeed(decryptedMnemonic)
-    const hdKey = this.sdk.keyPair.seedToHdKey(seed, network)
-    const privateKey = await this.deriveAddressPrivateKey(hdKey, source, network)
-
-    const recipient = this.sdk.keyPair.deriveShieldedAddress(seed, network, SHIELDED_ACCOUNT)
-    const senderOvk = this.sdk.keyPair.deriveShieldedOutgoingViewingKey(seed, network, SHIELDED_ACCOUNT)
-
-    const inputs = [new InputAddressWASM(source.platformAddress, source.nonce + 1, source.balanceCredits)]
-    const feeStrategy = [AddressFundsFeeStrategyStepWASM.DeductFromInput(0)]
-
-    const stateTransition = await this.sdk.shielded.createStateTransition('shield', {
-      recipient,
-      shieldAmount: amountCredits,
-      inputs,
-      privateKeys: [privateKey],
-      feeStrategy,
-      userFeeIncrease: 0,
-      memo: ShieldedMemoWASM.empty() as unknown as string,
-      senderOvk,
-    })
-
-    await this.sdk.stateTransitions.broadcast(stateTransition)
-    await this.sdk.stateTransitions.waitForStateTransitionResult(stateTransition)
-
-    return {
-      stHash: stateTransition.hash(false),
-      amountCredits: amountCredits.toString(),
-      fromAddress: source.platformAddress,
-    }
   }
 }
