+++ src/main/src/utils/index.ts
@@ -11,6 +11,7 @@ import * as migration0004 from '../../migrations/0004_pending_tx'
 import * as migration0005 from '../../migrations/0005_contacts'
 import * as migration0006 from '../../migrations/0006_asset_lock_fundings'
 import * as migration0007 from '../../migrations/0007_shielded_spent_notes'
+import * as migration0008 from '../../migrations/0008_platform_xpub'
 
 const migrations = [
   { name: '0000_init.ts', migration: migration0000 },
@@ -21,6 +22,7 @@ const migrations = [
   { name: '0005_contacts.ts', migration: migration0005 },
   { name: '0006_asset_lock_fundings.ts', migration: migration0006 },
   { name: '0007_shielded_spent_notes.ts', migration: migration0007 },
+  { name: '0008_platform_xpub.ts', migration: migration0008 },
 ]
 
 const inlineMigrationSource = {
